moved a couple of helper functions out of data builder and into helpers
